refactor: cli auth standalone page under _remote layout
Amp-Thread-ID: https://ampcode.com/threads/T-019d9e31-81dc-77a0-a9c4-1af7c1c36397
