+++ apps/connect/src/lib/config.ts
@@ -1,12 +1,10 @@
 import { getConnectors } from '@wagmi/core'
 import { Remote, Storage, TrustedHosts } from 'accounts'
 import { webAuthn } from 'accounts/wagmi'
-import { Address } from 'ox'
 import { http } from 'viem'
 import type { Capabilities } from 'viem/tempo'
 import { createConfig } from 'wagmi'
 import { tempo, tempoDevnet, tempoModerato } from 'wagmi/chains'
-import * as z from 'zod/mini'
 
 import * as Messenger from './messenger.js'
 
@@ -31,28 +29,6 @@ export const wagmiConfig = createConfig({
 
 export const chainIds = [tempo.id, tempoDevnet.id, tempoModerato.id] as const
 export type ChainId = (typeof chainIds)[number]
-
-export const zAddress = (options: { strict?: boolean } = {}) =>
-  z.string().check((ctx) => {
-    if (!Address.validate(ctx.value, options)) {
-      ctx.issues.push({
-        code: 'custom',
-        message: 'Invalid address',
-        input: ctx.value,
-      })
-    }
-  })
-
-export const zChainId = () =>
-  z.coerce.number().check((ctx) => {
-    if (!chainIds.some((id) => id === ctx.value)) {
-      ctx.issues.push({
-        code: 'custom',
-        input: ctx.value,
-        message: 'Unsupported chain ID',
-      })
-    }
-  })
 
 /** Remote context singleton. */
 export const remote = Remote.create({
