+++ apps/connect/src/routeTree.gen.ts
@@ -11,7 +11,6 @@
 import { Route as rootRouteImport } from './routes/__root'
 import { Route as EmailRouteImport } from './routes/email'
 import { Route as DesignRouteImport } from './routes/design'
-import { Route as CliRouteImport } from './routes/cli'
 import { Route as RemoteRouteImport } from './routes/_remote'
 import { Route as SplatRouteImport } from './routes/$'
 import { Route as RemoteIndexRouteImport } from './routes/_remote/index'
@@ -25,6 +24,7 @@ import { Route as RemoteRpcEth_signTypedData_v4RouteImport } from './routes/_rem
 import { Route as RemoteRpcEth_signTransactionRouteImport } from './routes/_remote/rpc/eth_signTransaction'
 import { Route as RemoteRpcEth_sendTransactionSyncRouteImport } from './routes/_remote/rpc/eth_sendTransactionSync'
 import { Route as RemoteRpcEth_sendTransactionRouteImport } from './routes/_remote/rpc/eth_sendTransaction'
+import { Route as RemoteAuthCliRouteImport } from './routes/_remote/auth.cli'
 
 const EmailRoute = EmailRouteImport.update({
   id: '/email',
@@ -34,11 +34,6 @@ const EmailRoute = EmailRouteImport.update({
 const DesignRoute = DesignRouteImport.update({
   id: '/design',
   path: '/design',
-  getParentRoute: () => rootRouteImport,
-} as any)
-const CliRoute = CliRouteImport.update({
-  id: '/cli',
-  path: '/cli',
   getParentRoute: () => rootRouteImport,
 } as any)
 const RemoteRoute = RemoteRouteImport.update({
@@ -111,14 +106,19 @@ const RemoteRpcEth_sendTransactionRoute =
     path: '/rpc/eth_sendTransaction',
     getParentRoute: () => RemoteRoute,
   } as any)
+const RemoteAuthCliRoute = RemoteAuthCliRouteImport.update({
+  id: '/auth/cli',
+  path: '/auth/cli',
+  getParentRoute: () => RemoteRoute,
+} as any)
 
 export interface FileRoutesByFullPath {
   '/$': typeof SplatRoute
   '/': typeof RemoteIndexRoute
-  '/cli': typeof CliRoute
   '/design': typeof DesignRoute
   '/email': typeof EmailRoute
   '/design/frames': typeof DesignFramesRoute
+  '/auth/cli': typeof RemoteAuthCliRoute
   '/rpc/eth_sendTransaction': typeof RemoteRpcEth_sendTransactionRoute
   '/rpc/eth_sendTransactionSync': typeof RemoteRpcEth_sendTransactionSyncRoute
   '/rpc/eth_signTransaction': typeof RemoteRpcEth_signTransactionRoute
@@ -131,11 +131,11 @@ export interface FileRoutesByFullPath {
 }
 export interface FileRoutesByTo {
   '/$': typeof SplatRoute
-  '/cli': typeof CliRoute
   '/design': typeof DesignRoute
   '/email': typeof EmailRoute
   '/design/frames': typeof DesignFramesRoute
   '/': typeof RemoteIndexRoute
+  '/auth/cli': typeof RemoteAuthCliRoute
   '/rpc/eth_sendTransaction': typeof RemoteRpcEth_sendTransactionRoute
   '/rpc/eth_sendTransactionSync': typeof RemoteRpcEth_sendTransactionSyncRoute
   '/rpc/eth_signTransaction': typeof RemoteRpcEth_signTransactionRoute
@@ -150,11 +150,11 @@ export interface FileRoutesById {
   __root__: typeof rootRouteImport
   '/$': typeof SplatRoute
   '/_remote': typeof RemoteRouteWithChildren
-  '/cli': typeof CliRoute
   '/design': typeof DesignRoute
   '/email': typeof EmailRoute
   '/design_/frames': typeof DesignFramesRoute
   '/_remote/': typeof RemoteIndexRoute
+  '/_remote/auth/cli': typeof RemoteAuthCliRoute
   '/_remote/rpc/eth_sendTransaction': typeof RemoteRpcEth_sendTransactionRoute
   '/_remote/rpc/eth_sendTransactionSync': typeof RemoteRpcEth_sendTransactionSyncRoute
   '/_remote/rpc/eth_signTransaction': typeof RemoteRpcEth_signTransactionRoute
@@ -170,10 +170,10 @@ export interface FileRouteTypes {
   fullPaths:
     | '/$'
     | '/'
-    | '/cli'
     | '/design'
     | '/email'
     | '/design/frames'
+    | '/auth/cli'
     | '/rpc/eth_sendTransaction'
     | '/rpc/eth_sendTransactionSync'
     | '/rpc/eth_signTransaction'
@@ -186,11 +186,11 @@ export interface FileRouteTypes {
   fileRoutesByTo: FileRoutesByTo
   to:
     | '/$'
-    | '/cli'
     | '/design'
     | '/email'
     | '/design/frames'
     | '/'
+    | '/auth/cli'
     | '/rpc/eth_sendTransaction'
     | '/rpc/eth_sendTransactionSync'
     | '/rpc/eth_signTransaction'
@@ -204,11 +204,11 @@ export interface FileRouteTypes {
     | '__root__'
     | '/$'
     | '/_remote'
-    | '/cli'
     | '/design'
     | '/email'
     | '/design_/frames'
     | '/_remote/'
+    | '/_remote/auth/cli'
     | '/_remote/rpc/eth_sendTransaction'
     | '/_remote/rpc/eth_sendTransactionSync'
     | '/_remote/rpc/eth_signTransaction'
@@ -223,7 +223,6 @@ export interface FileRouteTypes {
 export interface RootRouteChildren {
   SplatRoute: typeof SplatRoute
   RemoteRoute: typeof RemoteRouteWithChildren
-  CliRoute: typeof CliRoute
   DesignRoute: typeof DesignRoute
   EmailRoute: typeof EmailRoute
   DesignFramesRoute: typeof DesignFramesRoute
@@ -243,13 +242,6 @@ declare module '@tanstack/react-router' {
       path: '/design'
       fullPath: '/design'
       preLoaderRoute: typeof DesignRouteImport
-      parentRoute: typeof rootRouteImport
-    }
-    '/cli': {
-      id: '/cli'
-      path: '/cli'
-      fullPath: '/cli'
-      preLoaderRoute: typeof CliRouteImport
       parentRoute: typeof rootRouteImport
     }
     '/_remote': {
@@ -343,11 +335,19 @@ declare module '@tanstack/react-router' {
       preLoaderRoute: typeof RemoteRpcEth_sendTransactionRouteImport
       parentRoute: typeof RemoteRoute
     }
+    '/_remote/auth/cli': {
+      id: '/_remote/auth/cli'
+      path: '/auth/cli'
+      fullPath: '/auth/cli'
+      preLoaderRoute: typeof RemoteAuthCliRouteImport
+      parentRoute: typeof RemoteRoute
+    }
   }
 }
 
 interface RemoteRouteChildren {
   RemoteIndexRoute: typeof RemoteIndexRoute
+  RemoteAuthCliRoute: typeof RemoteAuthCliRoute
   RemoteRpcEth_sendTransactionRoute: typeof RemoteRpcEth_sendTransactionRoute
   RemoteRpcEth_sendTransactionSyncRoute: typeof RemoteRpcEth_sendTransactionSyncRoute
   RemoteRpcEth_signTransactionRoute: typeof RemoteRpcEth_signTransactionRoute
@@ -361,6 +361,7 @@ interface RemoteRouteChildren {
 
 const RemoteRouteChildren: RemoteRouteChildren = {
   RemoteIndexRoute: RemoteIndexRoute,
+  RemoteAuthCliRoute: RemoteAuthCliRoute,
   RemoteRpcEth_sendTransactionRoute: RemoteRpcEth_sendTransactionRoute,
   RemoteRpcEth_sendTransactionSyncRoute: RemoteRpcEth_sendTransactionSyncRoute,
   RemoteRpcEth_signTransactionRoute: RemoteRpcEth_signTransactionRoute,
@@ -379,7 +380,6 @@ const RemoteRouteWithChildren =
 const rootRouteChildren: RootRouteChildren = {
   SplatRoute: SplatRoute,
   RemoteRoute: RemoteRouteWithChildren,
-  CliRoute: CliRoute,
   DesignRoute: DesignRoute,
   EmailRoute: EmailRoute,
   DesignFramesRoute: DesignFramesRoute,
